chore(examples): adopt twd() plugin in tutorial and vue examples
tutorial-example: drop the dev-only initTWD block from main.tsx and
register twd() in vite.config.ts; alias twd-js/bundled to the local
src/dist copy since the tutorial does not install twd-js as a package.

vue-twd-example: same migration, plus relocate the dark theme literal
into vite.config.ts so initTWD options live in one place. main.ts is
now a 5-line Vue bootstrap.
